extract to utils
