+++ packages/cli/src/oclif/commands/run.js
@@ -6,7 +6,8 @@ const { callAPI } = require('../../utils/api');
 const { localAppCommand } = require('../../utils/local');
 const zapier = require('zapier-platform-core');
 const os = require('os');
-const fs = require('fs');
+
+const { parseAuth, parseInput } = require('../../utils/run');
 
 const METHODS = {
   perform: 'perform',
@@ -15,56 +16,6 @@ const METHODS = {
   test: 'test',
 };
 // const { listVersions } = require('../../utils/api');
-
-const authTypeFields = (type) => {
-  // another way for to find AuthOptions is by searching for the following phrases in keys:
-  const phrases = [
-    /.*api.*key.*/i,
-    /.*access.*token.*/i,
-    /.*token.*/i,
-    /.*user.*name.*/i,
-    /.*password.*/i,
-  ];
-  const authOptions = {};
-  phrases.forEach((phrase) => {
-    Object.keys(process.env).forEach((key) => {
-      if (key.match(phrase)) authOptions[key] = process.env[key];
-    });
-  });
-
-  return authOptions;
-};
-
-const parseAuth = (type, authInput) => {
-  // try to grab some sensible auth looking defaults - should really use the auth that is specified by the user
-  let auth = authTypeFields(type);
-
-  // override any of the environment variables with the ones provided via the command line
-  if (authInput) {
-    auth = { ...auth, ...JSON.parse(authInput) };
-  }
-  return auth;
-};
-
-const parseInput = (type, actionKey, input) => {
-  let inputParams = {};
-
-  // read the input from a json file if provided
-  const inputDataPath = path.join(process.cwd(), 'run-input.json');
-  if (fs.existsSync(inputDataPath)) {
-    const testInputData = require(inputDataPath);
-    if (type in testInputData && actionKey in testInputData[type]) {
-      inputParams = testInputData[type][actionKey];
-    }
-  }
-
-  // or provide JSON via the commandline, overriding any values from the run input file
-  if (input) {
-    inputParams = { ...inputParams, ...JSON.parse(input) };
-  }
-
-  return inputParams;
-};
 
 class RunCommand extends BaseCommand {
   async perform() {
@@ -100,6 +51,7 @@ class RunCommand extends BaseCommand {
       params: input,
     }; // do we need to pass in the auth? or we could just add it to the bundle afterwards
 
+    this.startSpinner('Retrieving bundle using input');
     // call the bundle API to construct the bundles
     // const app = await this.getWritableApp();
     const url = '/bundle'; // TODO set this to the url for the bundle API once that is implemented
@@ -117,7 +69,9 @@ class RunCommand extends BaseCommand {
     } catch (e) {
       console.log(e); // TODO handle errors
     }
+    this.stopSpinner();
 
+    this.startSpinner('Testing the provided method');
     // run the action using the provided bundle
     const appTester = zapier.createAppTester(App);
     let result;
@@ -131,9 +85,10 @@ class RunCommand extends BaseCommand {
           result = result.filter((field) => field.required === true);
         }
       } else {
-        // This is for runnig the Auth test if we implement that.
+        // This is for running the Auth test if we implement that.
         result = await appTester(App[actionType][method], bundle);
       }
+      this.stopSpinner();
     } catch (e) {
       console.log(e); // TODO handle errors
     }
@@ -167,8 +122,7 @@ RunCommand.flags = buildFlags({
 RunCommand.args = [
   {
     name: 'actionType',
-    description:
-      'The type of action (Only `creates` is currently supported). TODO: Add `triggers`, `searches`, `authentication`',
+    description: 'The type of action.', //  TODO: Test `authentication` to see if that works
     required: true,
   },
   {
